feat: recreating the code base and component architecture
