+++ package.json
@@ -21,6 +21,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-error-boundary": "^5.0.0",
+    "react-router-dom": "^7.6.2",
     "react-tree-graph": "^8.0.3"
   },
   "devDependencies": {
