+++ src/components/404Error.jsx
@@ -1,0 +1,5 @@
+export default function Error404(){
+    return (
+        <h1>404 Error Not Found ....</h1>
+    );
+}
