+++ src/components/home.jsx
@@ -1,0 +1,5 @@
+export default function Home() {
+    return (
+        <h1>Home - this is for the lading page</h1>
+    );
+}
