+++ src/components/modernView.jsx
@@ -1,0 +1,6 @@
+export default function ModernView(){
+    return (
+        <h1>Modern view - here inject the Sehansi's and Chanuka's works</h1>
+    );
+
+}
